+++ common/src/main/java/com/github/phylogeny/bettercushionplacement/network/GameRuleSyncPayload.java
@@ -2,16 +2,14 @@ package com.github.phylogeny.bettercushionplacement.network;
 
 import com.github.phylogeny.bettercushionplacement.Constants;
 import com.github.phylogeny.bettercushionplacement.registry.CommonGameRules;
-import com.github.phylogeny.bettercushionplacement.registry.RegistryObj;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.resources.Identifier;
 import net.minecraft.server.MinecraftServer;
-import net.minecraft.world.level.gamerules.GameRule;
 
-public record GameRuleSyncPayload(boolean allowInnerWallCushionPlacement)
+public record GameRuleSyncPayload(boolean allowInnerWallCushionPlacement, boolean cushionsSupportEachOther)
         implements CustomPacketPayload {
 
     public static final Identifier ID = Constants.MOD_IDENTIFIER.withPath("game_rule_sync");
@@ -22,15 +20,16 @@ public record GameRuleSyncPayload(boolean allowInnerWallCushionPlacement)
             StreamCodec.composite(
                     ByteBufCodecs.BOOL,
                     GameRuleSyncPayload::allowInnerWallCushionPlacement,
+                    ByteBufCodecs.BOOL,
+                    GameRuleSyncPayload::cushionsSupportEachOther,
                     GameRuleSyncPayload::new
             );
 
     public GameRuleSyncPayload(MinecraftServer server) {
-        this(getValue(server, CommonGameRules.ALLOW_INNER_WALL_CUSHION_PLACEMENT));
-    }
-
-    private static boolean getValue(MinecraftServer server, RegistryObj<GameRule<Boolean>> rule) {
-        return server.getGameRules().get(rule.get());
+        this(
+                CommonGameRules.ALLOW_INNER_WALL_CUSHION_PLACEMENT.get(server),
+                CommonGameRules.CUSHIONS_SUPPORT_EACH_OTHER.get(server)
+        );
     }
 
     @Override